runners: Add CommandOutputRunner for iterating a custom command
Add a new runner to iterate the execution of a custom command (which is
expected to produce a single line of numeric output for each execution),
and capture the timestamped values. The command execution is wrapped in
the cmd_iterate.sh script which adds separators and timestamps in the
same format as the other *iterate scripts.

Also add a simple test definition to exercise the runner.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>