testenv: Add get_env_parameter() function to read values from the environment
Add a get_env_parameter() equivalent to the get_test_parameter()
function exposed to test definitions. This can be used to read an
environment value with the same API as a test parameter (for instance,
to read sensitive values such as passwords that we don't want to store
as a test parameter).

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>