runners: Add MosquittoSubRunner for reading MQTT data
Add a new runner to capture MQTT data using the mosquitto_sub binary
distributed with the Mosquitto MQTT broker. The runner relies on the
JSON output format of mosquitto_sub, which in turn relies on the MQTT
payload being valid JSON. The runner supports subscribing to a topic and
capturing the values published, optionally supplying a JSON key to
extract from the payload.

To avoid passing MQTT password as a command line parameter, the runner
will instead write the --pw argument to a file in a temporary directory,
and overwrite XDG_CONFIG_HOME in the environment when running
mosquitto_sub, causing it to read the parameter from the config file.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>