Sync install docs content
README was missing Gentoo. Moved Arch linux up so the almost
alphabetically sorted linux distros would in fact be sorted.

I split OSX from "Other Linux" in index.html to match the other files. I
changed the content to point to intro.html#installing-flent as it wasn't
a one/two liner like the others so seemed to be better off as a link.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>